Changed blockStart() to shouldUpdate(). The processor can hold on to its sharedData until it decides to pull()
The RealTime's version of the sharedData can be accessed directly without having to make a copy of it since it can hold on to it until it decides to pull()

This is only useful for cases in which the processor has to depend on a snapshot of the GUI's sharedData between say, a keyswitch noteOn and a keyswitch noteOff.

If a GUI push is available when the keyswitch is active, the processor, at the start of the block should synthesize a keyswitch noteOff, pull the latest sharedData, then retrigger the same keyswitch note.